create halo logical regions

+++ src/LogicalUnstructured.cc
@@ -1,5 +1,5 @@
 /*
- * LogicalElement.cc
+ * LogicalUnstruc.cc
  *
  *  Created on: Sep 8, 2016
  *      Author: jgraham
@@ -25,11 +25,12 @@ LogicalUnstructured::LogicalUnstructured(Context ctx, HighLevelRuntime *runtime,
         PhysicalRegion pregion) :
     destroy_ispace(false),
     ispace(pregion.get_logical_region().get_index_space()),
+    ipartID(nullptr),
     destroy_fspace(false),
     fspace(pregion.get_logical_region().get_field_space()),
     destroy_lregion(false),
     lregion(pregion.get_logical_region()),
-    lpartID(NULL),
+    lpartID(nullptr),
     pregion(pregion),
     ctx(ctx),
     runtime(runtime)
@@ -45,78 +46,68 @@ LogicalUnstructured::LogicalUnstructured(Context ctx, HighLevelRuntime *runtime,
 
 LogicalUnstructured::LogicalUnstructured(Context ctx, HighLevelRuntime *runtime) :
     destroy_ispace(false),
-    ispaceID(NULL),
+    ispaceID(nullptr),
+    ipartID(nullptr),
     destroy_fspace(true),
     destroy_lregion(false),
-    lregionID(NULL),
-    lpartID(NULL),
+    lregionID(nullptr),
+    lpartID(nullptr),
     ctx(ctx),
     runtime(runtime)
 {
     fspace = runtime->create_field_space(ctx);
-    runtime->attach_name(fspace, "LogicalElement::fSpace");
+    runtime->attach_name(fspace, "LogicalUnstruc::fSpace");
     fspaceID = new FieldSpaceID;
     *fspaceID = fspace.get_id();
 }
 
 
 LogicalUnstructured::~LogicalUnstructured() {
-/*    if (lregionID != NULL) {
-        runtime->destroy_logical_region(ctx, lregion);
-        delete lregionID;
-    }
-    if (lpartID != NULL)
-        delete lpartID;
-    runtime->destroy_field_space(ctx, fspace);
-    delete fspaceID;
-    if (ispaceID != NULL) {
-        runtime->destroy_index_space(ctx, ispace);
-        delete ispaceID;
-    }
-    if (pregion.is_mapped())
-        runtime->unmap_region(ctx, pregion);
-
-*/
-/*    if (destroy_lregion)
+    if (destroy_lregion)
         runtime->destroy_logical_region(ctx, lregion);
     if (destroy_fspace)
         runtime->destroy_field_space(ctx, fspace);
     if (destroy_ispace)
         runtime->destroy_index_space(ctx, ispace);
-
-    if (lregionID != NULL)
-        delete lregionID;
-    if (lpartID != NULL)
-        delete lpartID;
-    delete fspaceID;
-    if (ispaceID != NULL)
-        delete ispaceID;*/
 }
 
 
 void LogicalUnstructured::partition(Coloring map, bool disjoint)
 {
-    assert( (ispaceID != NULL) && (lpartID == NULL) );
-    IndexPartition part = runtime->create_index_partition(ctx, ispace, map, disjoint);
-    runtime->attach_name(part, "LogicalElement::part");
-    lpart = runtime->get_logical_partition(ctx, lregion, part);
+    assert( (ispaceID != nullptr) && (lpartID == nullptr) );
+    ipart = runtime->create_index_partition(ctx, ispace, map, disjoint);
+    runtime->attach_name(ipart, "LogicalUnstruc::part");
+    ipartID = new IndexPartitionID;
+    *ipartID = ipart.id;
+    lpart = runtime->get_logical_partition(ctx, lregion, ipart);
     lpartID = new RegionTreeID;
     *lpartID = lpart.get_tree_id();
 }
 
 
-void LogicalUnstructured::allocate(int nElements)
+IndexSpace LogicalUnstructured::getSubspace(Color color)
 {
-    assert( (nElements > 0) && (fIDs.size() > 0) && (!pregion.is_mapped())
-            && (ispaceID == NULL) && (lregionID == NULL) );
+    assert( (ispaceID != nullptr) && (ipartID != nullptr) );
+    subspace = runtime->get_index_subspace(ctx, ipart, color);
+    char buf[43];
+    sprintf(buf, "LogicalUnstruc::iPart %d", color);
+    runtime->attach_name(subspace, buf);
+    return subspace;
+}
 
-    ispace = runtime->create_index_space(ctx, nElements);
+
+void LogicalUnstructured::allocate(int nUnstrucs)
+{
+    assert( (nUnstrucs > 0) && (fIDs.size() > 0) && (!pregion.is_mapped())
+            && (ispaceID == nullptr) && (lregionID == nullptr) );
+
+    ispace = runtime->create_index_space(ctx, nUnstrucs);
     destroy_ispace = true;
     char buf[43];
-    sprintf(buf, "LogicalElement::iSpace %d", nElements);
+    sprintf(buf, "LogicalUnstruc::iSpace %d", nUnstrucs);
     runtime->attach_name(ispace, buf);
     IndexAllocator allocator = runtime->create_index_allocator(ctx, ispace);
-    ptr_t begin = allocator.alloc(nElements);
+    ptr_t begin = allocator.alloc(nUnstrucs);
     assert(!begin.is_null());
     ispaceID = new IndexSpaceID;
     *ispaceID = ispace.get_id();
@@ -127,10 +118,10 @@ void LogicalUnstructured::allocate(int nElements)
 void LogicalUnstructured::allocate()
 {
     assert( (fIDs.size() > 0) && (!pregion.is_mapped())
-            && (ispaceID != NULL) && (lregionID == NULL) );
+            && (ispaceID != nullptr) && (lregionID == nullptr) );
     lregion = runtime->create_logical_region(ctx, ispace, fspace);
     destroy_lregion = true;
-    runtime->attach_name(lregion, "LogicalElement::lRegion");
+    runtime->attach_name(lregion, "LogicalUnstruc::lRegion");
     lregionID = new RegionTreeID;
     *lregionID = lregion.get_tree_id();
 }
@@ -138,7 +129,7 @@ void LogicalUnstructured::allocate()
 
 void LogicalUnstructured::addField(FieldID FID)
 {
-    assert(lregionID == NULL);
+    assert(lregionID == nullptr);
     fIDs.push_back(FID);
 }
 
@@ -171,7 +162,7 @@ void LogicalUnstructured::addField<int>(FieldID FID)
 
 PhysicalRegion LogicalUnstructured::getPRegion()
 {
-    assert(lregionID != NULL);
+    assert(lregionID != nullptr);
     if (!pregion.is_mapped() ) {
         RegionRequirement req(lregion, WRITE_DISCARD, EXCLUSIVE, lregion);
         for (int i=0; i<fIDs.size(); i++)
